+++ src/backend/ci/core/multijar/boot-multijar/src/main/kotlin/com/tencent/devops/multijar/MultijarApplication.kt
@@ -25,6 +25,25 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-subprojects {
-    group = "com.tencent.bk.devops.ci.mutijar"
+package com.tencent.devops.multijar
+
+import com.tencent.devops.common.service.MicroService
+import com.tencent.devops.common.service.MicroServiceApplication
+import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.FilterType
+
+@MicroService
+@ComponentScan(
+    "com.tencent.devops",
+    excludeFilters = [
+        ComponentScan.Filter(
+            type = FilterType.REGEX,
+            pattern = ["com\\.tencent\\.devops\\.common\\..*"]
+        )
+    ]
+)
+class MutijarApplication
+
+fun main(args: Array<String>) {
+    MicroServiceApplication.run(MutijarApplication::class, args)
 }
